move testHelpers.ts to testing export path
