+++ cmd/ui/src/views/Explore/ExploreSearch/ExploreSearch.test.tsx
@@ -20,8 +20,8 @@ import { act, render, screen } from 'src/test-utils';
 import ExploreSearch from './ExploreSearch';
 
 import userEvent from '@testing-library/user-event';
-import { encodeCypherQuery, mockCodemirrorLayoutMethods } from 'bh-shared-ui';
-import { createGraphKinds, cypherTestResponse } from 'bh-shared-ui/testing';
+import { encodeCypherQuery } from 'bh-shared-ui';
+import { createGraphKinds, cypherTestResponse, mockCodemirrorLayoutMethods } from 'bh-shared-ui/testing';
 import { ConfigurationKey, GraphData } from 'js-client-library';
 
 const comboboxLookaheadOptions = {
