+++ packages/javascript/bh-shared-ui/package.json
@@ -113,7 +113,7 @@
     },
     "exports": {
         ".": "./dist/index.js",
-        "./testing": "./dist/testing.js",
+        "./testing": "./dist/testing/index.js",
         "./FileIngest": "./dist/FileIngest.js",
         "./ApiExplorer": "./dist/ApiExplorer.js",
         "./SSOConfiguration": "./dist/SSOConfiguration.js",
