+++ packages/javascript/bh-shared-ui/src/components/CreateUserForm/CreateUserForm.test.tsx
@@ -17,7 +17,8 @@ import userEvent from '@testing-library/user-event';
 import { Dialog } from 'doodle-ui';
 import { MAX_EMAIL_LENGTH, MAX_NAME_LENGTH, MIN_NAME_LENGTH } from '../../constants';
 import { render, screen, waitFor } from '../../test-utils';
-import { PASSWORD_REQUIREMENTS_MESSAGE, setUpQueryClient } from '../../utils';
+import { setUpQueryClient } from '../../testing';
+import { PASSWORD_REQUIREMENTS_MESSAGE } from '../../utils';
 import { Roles } from '../../utils/roles';
 import CreateUserForm from './CreateUserForm';
