+++ packages/javascript/bh-shared-ui/src/testing/index.ts
@@ -13,4 +13,5 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
-export * from './mocks';
+export * from '../mocks';
+export * from './testHelpers';
