+++ packages/javascript/bh-shared-ui/src/utils/index.ts
@@ -42,7 +42,6 @@ export * from './quickUpload';
 export * from './roles';
 export * from './searchParams';
 export * from './strings';
-export * from './testHelpers';
 export * from './theme';
 export * from './type';
 export * from './user';
